+++ pyaerocom/__init__.py
@@ -57,4 +57,4 @@ from .utils import create_varinfo_table
 
 from . import aeroval
 
-from .data_access import download_minimal_dataset
+from .sample_data_access import download_minimal_dataset
