+++ pyaerocom/sample_data_access/minimal_dataset.py
