+++ tests/fixtures/data_access.py
@@ -19,7 +19,7 @@ from pyaerocom.io.readungriddedbase import ReadUngriddedBase
 logger = logging.getLogger(__name__)
 
 from pyaerocom import download_minimal_dataset
-from pyaerocom.data_access.minimal_dataset import minimal_dataset
+from pyaerocom.sample_data_access.minimal_dataset import minimal_dataset
 
 
 class DataForTests(NamedTuple):
